+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/actions/SearchAction.java
@@ -3,7 +3,6 @@ package org.lodder.subtools.multisubdownloader.actions;
 import java.util.List;
 
 import org.lodder.subtools.multisubdownloader.Messages;
-import org.lodder.subtools.multisubdownloader.gui.actions.ActionException;
 import org.lodder.subtools.multisubdownloader.gui.actions.search.SearchSetupException;
 import org.lodder.subtools.multisubdownloader.gui.dialog.Cancelable;
 import org.lodder.subtools.multisubdownloader.gui.dialog.progress.StatusListener;
